fix: Resolve Python 3.10 mypy errors and Windows viz tests
Python 3.10 numpy stubs are stricter than 3.13:
- Add explicit np.ndarray type annotations for shape-changing assignments
- Convert np.floating to float explicitly in changepoint.py
- Remove redundant casts in compare/base.py and multi_series.py

Windows CI fix:
- Set matplotlib backend to 'Agg' in conftest.py to avoid Tcl/Tk errors

+++ src/temporalcv/changepoint.py
@@ -474,7 +474,7 @@ def classify_regimes_from_changepoints(
             high_thresh = float(np.percentile(segment_values, 67))
         else:
             # Not enough segments - use overall statistics
-            overall_val = np.mean(segment_values)
+            overall_val = float(np.mean(segment_values))
             low_thresh = overall_val * 0.5
             high_thresh = overall_val * 1.5
